Update and rename ManhuaguiDTO.kt to Dto.kt
